Add info log statement to firmware example
This commit adds a single info level log statement to the firmware
example.

The motivation for this is that I think it might be helpful for users to
add their own logging when exploring/developing during the hackaton, and
with this change they won't have to think about importing `defmt::info`.